requirements.txt
pandas==1.5.3
numpy==1.23.5
opendatasets==0.1.22
Pillow==9.4.0
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
tqdm==4.65.0